Relocate io-related files
